Move all mutexes to a single top level mutex
This is a demonstration only. It removes the disk cache code as it's
inconvenient to convert to a mutexed version (it would have to
interact with the RPQ struct through the mutexed frontend). The
benchmark code does not include the disk cache anyway.

Before: insert: 494ms, receive: 397ms, total: 891ms
After:  insert: 296ms, receive: 157ms, total: 454ms